Fix test assertion for validate_pain function to use correct float syntax
- Updated the test case in `learn.rs` to replace the outdated float conversion with the new syntax `1e9_f32`, ensuring accurate validation of out-of-range pain scores.
- This change enhances the clarity and correctness of the test, aligning it with the latest Rust conventions.

+++ crates/dreamd-core/src/ingress/learn.rs
@@ -131,7 +131,7 @@ mod tests {
     #[test]
     fn rejects_out_of_range_required_scores() {
         assert_eq!(
-            LearnIngress::validate_pain(1e9 as f32).unwrap_err(),
+            LearnIngress::validate_pain(1e9_f32).unwrap_err(),
             LearnValidationError::PainOutOfRange
         );
         assert_eq!(
